+++ recipes/dotdeb.rb
@@ -1,12 +1,10 @@
 #
-# Author::  Joshua Timberman (<joshua@opscode.com>)
-# Author::  Seth Chisamore (<schisamo@opscode.com>)
 # Author::  Panagiotis Papadomitsos (<pj@ezgr.net>)
 #
 # Cookbook Name:: php
-# Recipe:: module_mysql
+# Recipe:: dotdeb
 #
-# Copyright 2009-2011, Opscode, Inc.
+# Copyright 2009-2012, Panagiotis Papadomitsos
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,12 +19,17 @@
 # limitations under the License.
 #
 
+if platform_family?('debian')
 
-pkg = value_for_platform_family(
-    [ 'rhel', 'fedora' ] => "php-#{node['php']['mysql_module_edition']}",
-    'debian' => "php5-#{node['php']['mysql_module_edition']}"
-)
+    apt_repository 'dotdeb-php' do
+      uri 'http://packages.dotdeb.org'
+      distribution node['php']['dotdeb_distribution']
+      components ['all']
+      key 'dotdeb.gpg'
+      notifies :run, 'execute[apt-get update]', :immediately
+      action :nothing
+      retries 2
+      retry_delay 2
+    end.run_action(:add)
 
-package pkg do
-  action :install
 end